[refactor] domain 패키지 이동
