+++ src/main/java/junseok/snr/wallet/api/controller/TransactionController.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.api.controller;
 
 import junseok.snr.wallet.api.controller.dto.WithdrawDto;
-import junseok.snr.wallet.api.domain.Transaction;
+import junseok.snr.wallet.domain.Transaction;
 import junseok.snr.wallet.api.service.TransactionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
