+++ src/main/java/junseok/snr/wallet/api/controller/dto/CreateWalletRequestDto.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.api.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
-import junseok.snr.wallet.api.domain.WalletType;
+import junseok.snr.wallet.domain.WalletType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
