--- src/main/java/junseok/snr/wallet/api/domain/WalletType.java
@@ -1,5 +1,0 @@
-package junseok.snr.wallet.api.domain;
-
-public enum WalletType {
-    ETH
-}
