+++ src/main/java/junseok/snr/wallet/api/repository/WalletRepository.java
@@ -1,6 +1,6 @@
 package junseok.snr.wallet.api.repository;
 
-import junseok.snr.wallet.api.domain.Wallet;
+import junseok.snr.wallet.domain.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WalletRepository extends JpaRepository<Wallet, Long> {
