+++ src/main/java/junseok/snr/wallet/api/service/EtherWalletService.java
@@ -3,8 +3,8 @@ package junseok.snr.wallet.api.service;
 import io.micrometer.common.util.StringUtils;
 import junseok.snr.wallet.common.Web3jUtils;
 import junseok.snr.wallet.api.controller.dto.CreateWalletRequestDto;
-import junseok.snr.wallet.api.domain.Wallet;
-import junseok.snr.wallet.api.domain.WalletType;
+import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.WalletType;
 import junseok.snr.wallet.api.repository.WalletRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
