+++ src/main/java/junseok/snr/wallet/api/service/TransactionService.java
@@ -3,9 +3,9 @@ package junseok.snr.wallet.api.service;
 import jakarta.persistence.criteria.Predicate;
 import junseok.snr.wallet.common.Web3jUtils;
 import junseok.snr.wallet.api.controller.dto.WithdrawDto;
-import junseok.snr.wallet.api.domain.Transaction;
-import junseok.snr.wallet.api.domain.TransactionType;
-import junseok.snr.wallet.api.domain.Wallet;
+import junseok.snr.wallet.domain.Transaction;
+import junseok.snr.wallet.domain.TransactionType;
+import junseok.snr.wallet.domain.Wallet;
 import junseok.snr.wallet.api.repository.TransactionRepository;
 import junseok.snr.wallet.api.repository.WalletRepository;
 import lombok.AllArgsConstructor;
