+++ src/main/java/junseok/snr/wallet/api/service/WalletService.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.api.service;
 
 import junseok.snr.wallet.api.controller.dto.CreateWalletRequestDto;
-import junseok.snr.wallet.api.domain.Wallet;
+import junseok.snr.wallet.domain.Wallet;
 
 import java.math.BigInteger;
