+++ src/main/java/junseok/snr/wallet/domain/TransactionStatus.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.api.domain;
+package junseok.snr.wallet.domain;
 
 public enum TransactionStatus {
     PENDING,
