+++ src/main/java/junseok/snr/wallet/domain/WalletType.java
@@ -1,0 +1,5 @@
+package junseok.snr.wallet.domain;
+
+public enum WalletType {
+    ETH
+}
