+++ src/test/java/junseok/snr/wallet/api/domain/TransactionTest.java
@@ -5,6 +5,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import junseok.snr.wallet.api.service.ExceptionCode;
 import junseok.snr.wallet.api.service.TransactionException;
 import junseok.snr.wallet.api.service.WalletException;
+import junseok.snr.wallet.domain.Transaction;
+import junseok.snr.wallet.domain.TransactionStatus;
+import junseok.snr.wallet.domain.TransactionType;
+import junseok.snr.wallet.domain.Wallet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
